--- src/assets/css/main.css
@@ -1,10 +1,0 @@
-.main-container {
-    display: flex;
-    flex-direction: column;
-    height: 100vh; /* 전체 화면 높이 */
-}
-
-router-view {
-    flex: 1; /* 나머지 공간을 모두 채우도록 설정 */
-    width: 100%; /* 전체 너비 */
-}
