--- src/components/Profile/Calendar/Month/Month.vue
@@ -1,488 +1,0 @@
-<template>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="initial-scale=1, width=device-width" />
-  </head>
-  <body>
-    <div class="div">
-      <main class="inner">
-        <section class="rectangle-parent">
-          <div class="frame-inner"></div>
-          <div class="weekday-labels">
-            <div class="weekday-labels-inner">
-              <div class="rectangle-group">
-                <div class="rectangle-div"></div>
-                <div class="day-bar-wrapper">
-                  <img
-                    class="day-bar-icon"
-                    loading="lazy"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/down.png"
-                  />
-                </div>
-                <div class="vertical-separator">2024</div>
-              </div>
-            </div>
-            <div class="weekday-labels-child">
-              <div class="frame-parent">
-                <div class="vector-wrapper">
-                  <img class="vector-icon" alt="" src="@/assets/icon/Profile/Calendar/Month/left.png" />
-                </div>
-                <div class="div3">4월</div>
-                <div class="vector-container">
-                  <img
-                    class="vector-icon1"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/right.png"
-                  />
-                </div>
-              </div>
-            </div>
-            <button class="rectangle-container">
-              <div class="frame-child1"></div>
-              <img
-                class="date-picker-icon"
-                alt=""
-                src="@/assets/icon/Profile/Calendar/Month/pencil.png"
-              />
-
-              <div class="div4">작성하기</div>
-            </button>
-          </div>
-          <div class="write-button">
-            <img
-              class="write-button-child"
-              alt=""
-              src="@/assets/icon/Profile/Calendar/Month/rectangle-19.png"
-            />
-
-            <div class="weekdays-wrapper">
-              <div class="weekdays">
-                <div class="month-headers">
-                  <img
-                    class="month-headers-child"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                  />
-
-                  <div class="vector-parent">
-                    <img
-                      class="rectangle-icon"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                    />
-
-                    <div class="div5">일</div>
-                  </div>
-                </div>
-                <div class="vector-group">
-                  <img
-                    class="frame-child2"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                  />
-
-                  <div class="div6">월</div>
-                </div>
-                <div class="group-div">
-                  <img
-                    class="frame-child3"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                  />
-
-                  <div class="div7">화</div>
-                </div>
-                <div class="vector-parent1">
-                  <img
-                    class="frame-child4"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                  />
-
-                  <div class="div8">수</div>
-                </div>
-                <div class="vector-parent2">
-                  <img
-                    class="frame-child5"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                  />
-
-                  <div class="div9">목</div>
-                </div>
-                <div class="vector-parent3">
-                  <img
-                    class="frame-child6"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                  />
-
-                  <div class="div10">금</div>
-                </div>
-                <div class="vector-parent4">
-                  <img
-                    class="frame-child7"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/circle.png"
-                  />
-
-                  <div class="div11">토</div>
-                </div>
-              </div>
-            </div>
-            <div class="months-and-days">
-              <div class="months-and-days-inner">
-                <div class="frame-group">
-                  <div class="vector-parent5">
-                    <img
-                      class="frame-child8"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectanglePink.png"
-                    />
-
-                    <div class="div12">7</div>
-                  </div>
-                  <div class="vector-parent6">
-                    <img
-                      class="frame-child9"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectanglePink.png"
-                    />
-
-                    <div class="div13">14</div>
-                  </div>
-                  <div class="vector-parent7">
-                    <img
-                      class="frame-child10"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectanglePink.png"
-                    />
-
-                    <div class="div14">21</div>
-                  </div>
-                  <div class="vector-parent8">
-                    <img
-                      class="frame-child11"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectanglePink.png"
-                    />
-
-                    <div class="div15">28</div>
-                  </div>
-                </div>
-              </div>
-              <div class="frame-container">
-                <div class="frame-div">
-                  <img
-                    class="frame-child12"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="vector-parent9">
-                    <img
-                      class="frame-child13"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                    />
-
-                    <div class="div16">1</div>
-                  </div>
-                </div>
-                <div class="vector-parent10">
-                  <img
-                    class="frame-child14"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="date-labels">8</div>
-                </div>
-                <div class="vector-parent11">
-                  <img
-                    class="frame-child15"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div17">15</div>
-                </div>
-                <div class="vector-parent12">
-                  <img
-                    class="frame-child16"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div18">22</div>
-                </div>
-                <div class="vector-parent13">
-                  <img
-                    class="frame-child17"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="main-page-calendar">29</div>
-                </div>
-              </div>
-              <div class="frame-parent1">
-                <div class="vector-parent14">
-                  <img
-                    class="frame-child18"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="vector-parent15">
-                    <img
-                      class="frame-child19"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                    />
-
-                    <div class="div19">2</div>
-                  </div>
-                </div>
-                <div class="vector-parent16">
-                  <img
-                    class="frame-child20"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div20">9</div>
-                </div>
-                <div class="vector-parent17">
-                  <img
-                    class="frame-child21"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div21">16</div>
-                </div>
-                <div class="vector-parent18">
-                  <img
-                    class="frame-child22"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div22">23</div>
-                </div>
-                <div class="vector-parent19">
-                  <img
-                    class="frame-child23"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div23">30</div>
-                </div>
-              </div>
-              <div class="frame-parent2">
-                <div class="vector-parent20">
-                  <img
-                    class="frame-child24"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="vector-parent21">
-                    <img
-                      class="frame-child25"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                    />
-
-                    <div class="div24">3</div>
-                  </div>
-                </div>
-                <div class="vector-parent22">
-                  <img
-                    class="frame-child26"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div25">10</div>
-                </div>
-                <div class="vector-parent23">
-                  <img
-                    class="frame-child27"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div26">17</div>
-                </div>
-                <div class="vector-parent24">
-                  <img
-                    class="frame-child28"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div27">24</div>
-                </div>
-              </div>
-              <div class="frame-parent3">
-                <div class="vector-parent25">
-                  <img
-                    class="frame-child29"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="vector-parent26">
-                    <img
-                      class="frame-child30"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                    />
-
-                    <div class="div28">4</div>
-                  </div>
-                </div>
-                <div class="vector-parent27">
-                  <img
-                    class="frame-child31"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div29">11</div>
-                </div>
-                <div class="vector-parent28">
-                  <img
-                    class="frame-child32"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div30">18</div>
-                </div>
-                <div class="vector-parent29">
-                  <img
-                    class="frame-child33"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div31">25</div>
-                </div>
-              </div>
-              <div class="frame-parent4">
-                <div class="vector-parent30">
-                  <img
-                    class="frame-child34"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="vector-parent31">
-                    <img
-                      class="frame-child35"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                    />
-
-                    <div class="div32">5</div>
-                  </div>
-                </div>
-                <div class="vector-parent32">
-                  <img
-                    class="frame-child36"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div33">12</div>
-                </div>
-                <div class="vector-parent33">
-                  <img
-                    class="frame-child37"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div34">19</div>
-                </div>
-                <div class="vector-parent34">
-                  <img
-                    class="frame-child38"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleSkin.png"
-                  />
-
-                  <div class="div35">26</div>
-                </div>
-              </div>
-              <div class="frame-parent5">
-                <div class="vector-parent35">
-                  <img
-                    class="frame-child39"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleBlue.png"
-                  />
-
-                  <div class="vector-parent36">
-                    <img
-                      class="frame-child40"
-                      alt=""
-                      src="@/assets/icon/Profile/Calendar/Month/rectangleBlue.png"
-                    />
-
-                    <div class="div36">6</div>
-                  </div>
-                </div>
-                <div class="vector-parent37">
-                  <img
-                    class="frame-child41"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleBlue.png"
-                  />
-
-                  <div class="div37">13</div>
-                </div>
-                <div class="vector-parent38">
-                  <img
-                    class="frame-child42"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleBlue.png"
-                  />
-
-                  <div class="div38">20</div>
-                </div>
-                <div class="vector-parent39">
-                  <img
-                    class="frame-child43"
-                    alt=""
-                    src="@/assets/icon/Profile/Calendar/Month/rectangleBlue.png"
-                  />
-
-                  <div class="div39">27</div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-      </main>
-    </div>
-  </body>
-</html>
-
-</template>
-
-<script setup>
-
-</script>
-
-<style>
-@import "@/assets/css/Profile/Calendar/Month/global.css";
-@import "@/assets/css/Profile/Calendar/Month/Month.css";
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
-</style>
