--- src/components/Profile/Statistics/BlockGraph/BlockGraph.vue
@@ -1,101 +1,0 @@
-<template>
-    <div class="div">
-      <section class="inner">
-        <div class="frame-parent">
-          <div class="frame-wrapper">
-            <div class="rectangle-parent">
-              <div class="frame-item"></div>
-              <b class="b">영화 장르 랭킹</b>
-              <div class="vector-wrapper">
-                <img class="vector-icon" alt="" src="./public/vector.svg" />
-              </div>
-            </div>
-          </div>
-          <div class="frame-group">
-            <div class="frame-container">
-              <div class="vector-parent">
-                <img class="vector-icon1" alt="" src="./public/vector-1.svg" />
-
-                <div class="vector-container">
-                  <img
-                    class="vector-icon2"
-                    alt=""
-                    src="./public/vector-2.svg"
-                  />
-                </div>
-              </div>
-              <div class="wrapper">
-                <b class="b1">30</b>
-              </div>
-              <div class="container">
-                <b class="b2">20</b>
-              </div>
-              <div class="frame">
-                <b class="b3">15</b>
-              </div>
-              <div class="frame-div">
-                <b class="b4">10</b>
-              </div>
-              <div class="wrapper1">
-                <b class="b5">5</b>
-              </div>
-              <img class="vector-icon3" alt="" src="./public/vector-3.svg" />
-            </div>
-            <div class="frame-wrapper1">
-              <div class="frame-parent1">
-                <div class="rectangle-group">
-                  <div class="frame-inner"></div>
-                  <div class="rectangle-div"></div>
-                  <div class="frame-child1"></div>
-                  <div class="wrapper-arrow-2">
-                    <img
-                      class="wrapper-arrow-2-child"
-                      loading="lazy"
-                      alt=""
-                      src="./public/arrow-2.svg"
-                    />
-                  </div>
-                  <img
-                    class="arrow-icon"
-                    loading="lazy"
-                    alt=""
-                    src="./public/arrow-3.svg"
-                  />
-                </div>
-                <div class="frame-parent2">
-                  <div class="frame-wrapper2">
-                    <div class="frame-parent3">
-                      <button class="rectangle-container">
-                        <div class="frame-child2"></div>
-                        <b class="b6">코미디</b>
-                      </button>
-                      <button class="group-button">
-                        <div class="frame-child3"></div>
-                        <b class="b7">스릴러</b>
-                      </button>
-                      <button class="rectangle-parent1">
-                        <div class="frame-child4"></div>
-                        <b class="b8">드라마</b>
-                      </button>
-                    </div>
-                  </div>
-                  <b class="konan-container">
-                    <span class="span1">‘스릴러’</span>
-                      <span> 장르를 가장 선호하시는군요!</span>
-                  </b>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-    </div>
-</template>
-
-<script setup>
-
-</script>
-
-<style scoped>
-
-</style>
