--- src/components/Profile/Statistics/LineGraph/LineGraph.vue
@@ -1,324 +1,0 @@
-<template>
-    <div class="div">
-      <section class="button-set">
-        <div class="progress-bar">
-          <div class="slider">
-            <div class="table-creator">
-              <div class="frame-parent">
-                <div class="icon-library-parent">
-                  <button class="icon-library">
-                    <div class="icon-library-child"></div>
-                    <b class="b">연 간</b>
-                  </button>
-                  <button class="icon-library1">
-                    <div class="icon-library-item"></div>
-                    <b class="b1">월 간</b>
-                  </button>
-                </div>
-                <div class="square-shape">
-                  <div class="rectangle-parent">
-                    <div class="frame-child"></div>
-                    <div class="triangle-shape">
-                      <b class="b2">2024년</b>
-                    </div>
-                    <b class="b3">4월</b>
-                    <div class="oval-shape">
-                      <div class="path-creator-parent">
-                        <div class="path-creator"></div>
-                        <img
-                          class="vector-icon"
-                          loading="lazy"
-                          alt=""
-                          src="./public/vector.svg"
-                        />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div class="aligner">
-              <div class="distributer">
-                <div class="stacker">
-                  <div class="rectangle-group">
-                    <div class="frame-item"></div>
-                    <b class="b4">
-                      <p class="p">감</p>
-                      <p class="p1">정</p>
-                    </b>
-                    <b class="b5">
-                      <p class="p2">그</p>
-                      <p class="p3">래</p>
-                      <p class="p4">프</p>
-                    </b>
-                  </div>
-                </div>
-                <div class="anchor-point">
-                  <div class="anchor-point-child"></div>
-                  <div class="row-manager">
-                    <div class="matrix-transform-parent">
-                      <div class="matrix-transform">
-                        <div class="scale-control">
-                          <div class="rotate-control">
-                            <img
-                              class="icon"
-                              loading="lazy"
-                              alt=""
-                              src="./public/-20240405--317@2x.png"
-                            />
-                          </div>
-                          <div class="skew-control">
-                            <div class="blend-mode-manager">
-                              <img
-                                class="icon1"
-                                loading="lazy"
-                                alt=""
-                                src="./public/-20240405--402@2x.png"
-                              />
-
-                              <div class="fill-color-picker">
-                                <div class="stroke-color-picker">9</div>
-                              </div>
-                            </div>
-                            <div class="background-filler">
-                              <div class="layer-orderer">
-                                <div class="data-aggregator">
-                                  <img
-                                    class="icon2"
-                                    loading="lazy"
-                                    alt=""
-                                    src="./public/-20240405--318@2x.png"
-                                  />
-                                </div>
-                                <div class="input-combiner">7</div>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                        <div class="value-comparator">
-                          <img
-                            class="icon3"
-                            loading="lazy"
-                            alt=""
-                            src="./public/-20240405--405@2x.png"
-                          />
-
-                          <div class="loop-manager">
-                            <div class="div3">5</div>
-                          </div>
-                        </div>
-                        <div class="edge-connection">
-                          <div class="subroutine-container">
-                            <div class="variable-assigner">
-                              <img
-                                class="icon4"
-                                alt=""
-                                src="./public/-20240405--317-1@2x.png"
-                              />
-                            </div>
-                            <div class="operation-processor">3</div>
-                          </div>
-                          <div class="expression-evaluator">
-                            <div class="decision-maker">
-                              <img
-                                class="icon5"
-                                alt=""
-                                src="./public/-20240405--405-1@2x.png"
-                              />
-
-                              <div class="parallel-processor">
-                                <div class="nested-loop">1</div>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                      <div class="frame-inner"></div>
-                    </div>
-                    <div class="matrix-manipulator">
-                      <div class="div4">4/1</div>
-                    </div>
-                  </div>
-                  <div class="ruler-parent">
-                    <div class="ruler">
-                      <div class="grid-creator"></div>
-                    </div>
-                    <div class="column-manager">4/7</div>
-                  </div>
-                  <div class="frame-group">
-                    <div class="line-wrapper">
-                      <div class="line-div"></div>
-                    </div>
-                    <div class="div5">4/14</div>
-                  </div>
-                  <div class="frame-container">
-                    <div class="line-container">
-                      <div class="frame-child1"></div>
-                    </div>
-                    <div class="div6">4/21</div>
-                  </div>
-                  <div class="frame-div">
-                    <div class="line-frame">
-                      <div class="frame-child2"></div>
-                    </div>
-                    <div class="div7">4/28</div>
-                  </div>
-                  <div class="anchor-point-item"></div>
-                  <div class="anchor-point-inner"></div>
-                  <div class="ellipse-div"></div>
-                  <div class="anchor-point-child1"></div>
-                  <img
-                    class="ellipse-icon"
-                    loading="lazy"
-                    alt=""
-                    src="./public/ellipse-5.svg"
-                  />
-
-                  <div class="anchor-point-child2"></div>
-                  <div class="index-tracker">
-                    <div class="key-value-pair">
-                      <div class="list-processor">
-                        <div class="stack-manager"></div>
-                      </div>
-                      <img
-                        class="queue-handler-icon"
-                        alt=""
-                        src="./public/vector-1.svg"
-                      />
-                    </div>
-                  </div>
-                </div>
-              </div>
-              <div class="priority-queue">
-                <div class="hash-table">
-                  <div class="rectangle-container">
-                    <div class="rectangle-div"></div>
-                    <b class="b6">
-                      <p class="p5">감</p>
-                      <p class="p6">정</p>
-                    </b>
-                    <b class="b7">
-                      <p class="p7">랭</p>
-                      <p class="p8">킹</p>
-                    </b>
-                  </div>
-                </div>
-                <div class="depth-first-search">
-                  <div class="depth-first-search-child"></div>
-                  <div class="graph-builder">
-                    <div class="path-finder">
-                      <img
-                        class="icon6"
-                        alt=""
-                        src="./public/-20240405--317-2@2x.png"
-                      />
-
-                      <div class="max-heap">
-                        <img
-                          class="icon7"
-                          alt=""
-                          src="./public/-20240405--318-1@2x.png"
-                        />
-                      </div>
-                      <img
-                        class="icon8"
-                        alt=""
-                        src="./public/-20240405--317-3@2x.png"
-                      />
-                    </div>
-                  </div>
-                  <div class="b-search-tree">
-                    <div class="red-black-tree">
-                      <div class="red-black-tree-child"></div>
-                      <div class="trie">
-                        <img
-                          class="group-icon2"
-                          alt=""
-                          src="./public/group-2.svg"
-                        />
-                      </div>
-                      <img
-                        class="icon9"
-                        alt=""
-                        src="./public/-20240405--402-1@2x.png"
-                      />
-
-                      <div class="zip-algorithm">
-                        <div class="deflate-algorithm">
-                          <div class="gzip-algorithm">
-                            <div class="div8">좋음이</div>
-                          </div>
-                          <b class="b8">15회</b>
-                        </div>
-                      </div>
-                    </div>
-                    <div class="red-black-tree1">
-                      <div class="red-black-tree-item"></div>
-                      <div class="group-wrapper">
-                        <img
-                          class="group-icon3"
-                          alt=""
-                          src="./public/group-3.svg"
-                        />
-                      </div>
-                      <img
-                        class="icon10"
-                        alt=""
-                        src="./public/-20240405--405-2@2x.png"
-                      />
-
-                      <div class="red-black-tree-inner">
-                        <div class="parent">
-                          <div class="div9">보통이</div>
-                          <b class="b9">10회</b>
-                        </div>
-                      </div>
-                    </div>
-                    <div class="red-black-tree2">
-                      <div class="red-black-tree-child1"></div>
-                      <div class="group-container">
-                        <img
-                          class="group-icon4"
-                          alt=""
-                          src="./public/group-4.svg"
-                        />
-                      </div>
-                      <div class="group">
-                        <img
-                          class="icon11"
-                          alt=""
-                          src="./public/-20240405--405-3@2x.png"
-                        />
-
-                        <div class="levenshtein-distance">
-                          <div class="div10">나쁨이</div>
-                        </div>
-                      </div>
-                      <div class="wrapper">
-                        <b class="b10">6회</b>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div class="run-length-encoding">
-            <b class="b11">
-              <span>이번 달은 </span>
-              <span class="span">좋음이</span>
-              <span>가 많은 날들을 경험하셨군요! </span>
-            </b>
-          </div>
-        </div>
-      </section>
-    </div>
-</template>
-
-<script setup>
-
-</script>
-
-<style scoped>
-
-</style>
