--- src/components/Profile/Statistics/Statistics.vue
@@ -1,8 +1,0 @@
-<template>
-  <LineGraph/>
-  <BlockGraph/>
-</template>
-<script setup>
-import LineGraph from "@/components/Profile/Statistics/LineGraph/LineGraph.vue";
-import BlockGraph from "@/components/Profile/Statistics/BlockGraph/BlockGraph.vue";
-</script>
