+++ src/components/common/MainHeader.vue
@@ -138,4 +138,3 @@ const logout = async () => {
 }
 
 </style>
-
