+++ src/components/header/Mypage.vue
@@ -1,128 +1,156 @@
 <template>
   <div class="div">
-    <section class="inner">
+    <img class="child" alt="" src="@/assets/icon/MyPage/blurRectangle.png" />
+
+    <b class="b">내 정보</b>
+    <b class="b1">내 다이어리</b>
+    <b class="b2">개발중</b>
+    <b class="b3">내 관심 랭킹</b>
+    <img class="vector-icon" alt="" src="@/assets/icon/MyPage/blueCategory.png" />
+    <div class="inner">
       <div class="frame-parent">
-        <div class="frame-wrapper">
-          <div class="frame-group">
-            <div class="konan-wrapper">
-              <h1 class="konan">Konan님</h1>
-            </div>
-            <h2 class="h2">오늘 하루 기분은 어떠신가요?</h2>
-          </div>
+        <div class="konan-wrapper">
+          <h1 class="konan">Konan님</h1>
         </div>
-
-        <div class="frame-container">
-          <div class="frame-div">
-            <div class="frame-parent1">
-              <div class="vector-parent">
-                <img
-                    class="frame-inner"
-                    alt=""
-                    src="@/assets/icon/Profile/Info/my.png"
-                />
-
-                <b class="b">내 정보</b>
-                <div class="vector-wrapper">
-                  <img
-                      class="vector-icon"
-                      loading="lazy"
-                      alt=""
-                      src="@/assets/icon/Profile/Info/my.png"
-                  />
-                </div>
-              </div>
-
-
-              <div class="vector-group">
-                <img
-                    class="rectangle-icon"
-                    alt=""
-                    src="@/assets/icon/Profile/Info/rectangle-16.png"
-                />
-                <b class="b1">내 관심 랭킹</b>
-                <div class="vector-container">
-                  <img
-                      class="vector-icon1"
-                      alt=""
-                      src="@/assets/icon/Profile/Info/heartLine.png"
-                  />
-                </div>
-              </div>
-
+        <div class="frame-wrapper">
+          <div class="wrapper-rectangle-16-parent">
+            <div class="wrapper-rectangle-16">
+              <img
+                  class="wrapper-rectangle-16-child"
+                  alt=""
+                  src="@/assets/icon/MyPage/blurRectangle.png"
+              />
             </div>
-            <div class="frame-parent2">
-              <div class="group-div">
+            <div class="wrapper-rectangle-18-parent">
+              <div class="wrapper-rectangle-18">
                 <img
-                    class="frame-child1"
+                    class="wrapper-rectangle-18-child"
                     alt=""
-                    src="@/assets/icon/Profile/Info/rectangle-16.png"
+                    src="@/assets/icon/MyPage/blurRectangle.png"
                 />
+              </div>
+              <b class="ai">AI 추천 목록</b>
+            </div>
+            <img
+                class="image-2-icon"
+                loading="lazy"
+                alt=""
+                src="@/assets/icon/MyPage/AI.png"
+            />
 
-                <b class="b2">내 캘린더</b>
-                <div class="group-wrapper">
+            <div class="wrapper-rectangle-19-parent">
+              <div class="wrapper-rectangle-19">
+                <img
+                    class="wrapper-rectangle-19-child"
+                    alt=""
+                    src="@/assets/icon/MyPage/blurRectangle.png"
+                />
+              </div>
+              <b class="b4">내 캘린더</b>
+              <div class="wrapper-group-wrapper">
+                <div class="wrapper-group">
                   <img
                       class="group-icon2"
                       alt=""
-                      src="@/assets/icon/Profile/Info/calendar.png"
-                  />
-                </div>
-              </div>
-              <div class="vector-parent1">
-                <img
-                    class="frame-child2"
-                    alt=""
-                    src="@/assets/icon/Profile/Info/rectangle-16.png"
-                />
-
-                <b class="b3">내 다이어리</b>
-                <div class="group-container">
-                  <img
-                      class="group-icon3"
-                      alt=""
-                      src="@/assets/icon/Profile/Info/diary.png"
+                      src="@/assets/icon/MyPage/calendar.png"
                   />
                 </div>
               </div>
             </div>
-          </div>
-          <div class="frame-parent3">
-            <div class="vector-parent2">
-              <img
-                  class="frame-child3"
-                  alt=""
-                  src="@/assets/icon/Profile/Info/rectangle-16.png"
-              />
-
-              <b class="ai">AI 추천 목록</b>
-              <div class="image-2-wrapper">
+            <div class="wrapper-rectangle-20-parent">
+              <div class="wrapper-rectangle-20">
                 <img
-                    class="image-2-icon"
+                    class="wrapper-rectangle-20-child"
+                    alt=""
+                    src="@/assets/icon/MyPage/blurRectangle.png"
+                />
+              </div>
+              <div class="wrapper-group1">
+                <img class="group-icon3" alt="" src="@/assets/icon/MyPage/diary.png" />
+              </div>
+            </div>
+            <div class="wrapper-rectangle-21-parent">
+              <div class="wrapper-rectangle-21">
+                <img
+                    class="wrapper-rectangle-21-child"
+                    alt=""
+                    src="@/assets/icon/MyPage/blurRectangle.png"
+                />
+              </div>
+              <div class="wrapper-vector">
+                <img
+                    class="vector-icon1"
                     loading="lazy"
                     alt=""
-                    src="@/assets/icon/Profile/Info/ai.png"
+                    src="@/assets/icon/MyPage/blueCategory.png"
                 />
               </div>
             </div>
-            <div class="vector-parent3">
+            <div class="wrapper-vector1">
+              <img class="vector-icon2" alt="" src="@/assets/icon/MyPage/my.png" />
+            </div>
+            <h2 class="h2">오늘 하루 기분은 어떠신가요?</h2>
+            <div class="vector-parent">
               <img
-                  class="frame-child4"
+                  class="frame-inner"
                   alt=""
-                  src="@/assets/icon/Profile/Info/rectangle-16.png"
+                  src="@/assets/icon/MyPage/rectangle.png"
               />
 
-              <b class="b4">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;개발중</b>
-              <div class="vector-frame">
+              <div class="vector-group">
                 <img
-                    class="vector-icon2"
+                    class="vector-icon3"
                     alt=""
-                    src="@/assets/icon/Profile/Info/categoryBlue.png"
+                    src="@/assets/icon/MyPage/x.png"
                 />
+
+                <div class="wrapper">
+                  <h1 class="h1">로그인 정보</h1>
+                </div>
+              </div>
+              <div class="frame-container">
+                <div class="frame-group">
+                  <div class="container">
+                    <div class="div3">닉네임</div>
+                  </div>
+                  <div class="frame-div">
+                    <div class="konan-parent">
+                      <b class="konan1">Konan</b>
+                      <div class="frame">
+                        <div class="div4">변경하기</div>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="line-div"></div>
+                </div>
+              </div>
+              <div class="frame-wrapper1">
+                <div class="frame-parent1">
+                  <div class="wrapper1">
+                    <div class="div5">연결된 계정</div>
+                  </div>
+                  <div class="frame-parent2">
+                    <img
+                        class="frame-child1"
+                        loading="lazy"
+                        alt=""
+                        src="@/assets/icon/MyPage/google.png"
+                    />
+
+                    <div class="dlwldms0902gmailcom-wrapper">
+                      <b class="dlwldms0902gmailcom">dlwldms0902@gmail.com</b>
+                    </div>
+                  </div>
+                  <div class="wrapper2">
+                    <div class="div6">변경하기</div>
+                  </div>
+                </div>
               </div>
             </div>
           </div>
         </div>
       </div>
-    </section>
+    </div>
   </div>
 </template>
 
@@ -131,7 +159,7 @@
 </script>
 
 <style>
-@import "@/assets/css/Profile/Info/global.css";
-@import "@/assets/css/Profile/Info/Mypage.css";
+@import "@/assets/css/MyPage/MyPage.css";
+@import "@/assets/css/MyPage/global.css";
 @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
 </style>
