+++ src/components/header/statistics/Movie-BarGraph.vue
