--- src/components/main/Welcome/Welcome.vue
@@ -1,56 +1,0 @@
-<template>
-    <div class="div">
-        <header class="frame-parent"></header>
-        <section class="inner">
-        <div class="rectangle-parent">
-            <div class="frame-item"></div>
-            <div class="frame-group">
-            <div class="frame-wrapper">
-                <div class="message-box-parent">
-                <div class="message-box">
-                    <img
-                    class="icon"
-                    loading="lazy"
-                    alt=""
-                    src="@/assets/icon/MainPage/Welcome/-20240405--317@2x.png"/>
-                </div>
-                <div class="frame-container">
-                    <div class="konan-wrapper">
-                    <h2 class="konan">Konan 님</h2>
-                    </div>
-                    <div class="div1">가입을 진심으로 축하드립니다!</div>
-                </div>
-                </div>
-            </div>
-            <div class="line-wrapper">
-                <div class="frame-inner"></div>
-            </div>
-            <div class="frame-div">
-                <div class="parent">
-                <div class="div2">오늘의 기분을 소중한 기억을</div>
-                <div class="wrapper">
-                    <div class="div3">다이어리에 담아보세요!</div>
-                </div>
-                </div>
-            </div>
-            </div>
-            <div class="frame-wrapper1">
-            <button class="rectangle-group">
-                <div class="rectangle-div"></div>
-                <b class="b">가입 완료</b>
-            </button>
-            </div>
-        </div>
-        </section>
-    </div>
-</template>
-
-<script setup>
-
-</script>
-
-<style>
-@import '@/assets/css/MainPage/Welcome/global.css';
-@import '@/assets/css/MainPage/Welcome/Welcome.css';
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
-</style>
