+++ src/components/mypage/AIrecommended/Card/Card.vue
