+++ src/components/mypage/AIrecommended/List/List.vue
